Update 10-vbr.conf
The Source of the logs in system messages is coming from python3. The rsyslog variable $programname is to match the application that created the log, or the Source; so the honeypots were not being matched. Instead, the honeypots are contained within the message, so I Replaced $programname startswith with $msg contains so the honeypots are matched. This might be able to be condensed further with matching *_honeypot, but I didn't want to go too far and match too much.

Additionally, added . to the then statement to tell rsyslog to forward the matched message with its original facility/severity to the remote server.

And a stop to not log the message anywhere else to avoid duplicates.

As stated in VeeamHub#3 this fixes the logs being forwarded to a remote EventLog Analyzer server.